+++ tilings/tilescope.py
@@ -140,10 +140,13 @@ class GuidedSearcher(TileScope):
         *args,
         **kwargs
     ):
-        self.tilings = frozenset(tilings)
+        self.tilings = frozenset(t.remove_assumptions() for t in tilings)
         super().__init__(basis, pack, *args, **kwargs)
         for t in self.tilings:
-            self._add_to_queue(self.classdb.get_label(t))
+            class_label = self.classdb.get_label(t)
+            is_empty = self.classdb.is_empty(t, class_label)
+            if not is_empty:
+                self._add_to_queue(class_label)
 
     def _expand(
         self,